Set accept_ra to 2 when enabling IPv6.
The default value of accept_ra causes the kernel to accept
Router Advertisements only if forwarding is off. This causes
problems with clatd, which uses IPv6 forwarding on a tun
interface, and it may cause problems with IPv6 tethering when
we implement it.

Fix this by setting accept_ra to 2, which accepts RAs regardless
of whether forwarding is on.

Bug: 8276725
Change-Id: Ifee0ceaa68ce438a845fa2d7f3b76c60857114a0