Temporarily revert the "talk to netlink directly" change.
The change being reverted is http://ag/486277. Instead, use /sbin/ip again. The
code to talk to netlink fails on volantis. I.e., instead of this:
    $ ip route show table 1006
    default via 100.110.191.254 dev wlan0
    100.110.128.0/18 dev wlan0  scope link
we end up with this:
    $ ip route show table 1006
    default dev wlan0  proto static
    100.110.128.0/18 dev wlan0  proto static
Notice the lack of the nexthop and the addition of "proto static". I think the
netlink message is somehow not properly aligned on volantis, causing the kernel
to misinterpret it.

Bug: 15840054
Change-Id: Ief60473e337410f7cb35890de0a5a74a21723a41